+++ xarray/tests/test_backends.py
@@ -391,7 +391,8 @@ class CFEncodedDataTest(DatasetIOTestCases):
         # make sure roundtrip encoding didn't change the
         # original dataset.
         self.assertDatasetIdentical(encoded,
-                                    create_encoded_unsigned_masked_scaled_data()
+                                    create_encoded_unsigned_masked_scaled_data(
+                                    )
                                     )
         with self.roundtrip(encoded) as actual:
             self.assertDatasetAllClose(decoded, actual)